database setup, DRF Authentication , Token Based Authentication

+++ api/views.py
@@ -1,34 +1,39 @@
-from django.shortcuts import render
 from rest_framework.response import Response
 from rest_framework.decorators import api_view
-from .serializers import TaskSerializers
+from rest_framework import status
 from .models import Task
-
+from .serializers import TaskSerializer
 
 @api_view(['GET'])
-def getTasks(request):
+def getTask(request):
     tasks = Task.objects.all()
-    serializer = TaskSerializers(tasks, many=True)
-    return Response(serializer.data)
+    serializer = TaskSerializer(tasks, many=True)
+    return Response(serializer.data, status=status.HTTP_200_OK)
 
 @api_view(['POST'])
-def createTasks(request):
-    serializer = TaskSerializers(data=request.data)
+def createTask(request):
+    serializer = TaskSerializer(data=request.data)
     if serializer.is_valid():
         serializer.save()
-    return Response(serializer.data)
+        return Response(serializer.data, status=status.HTTP_201_CREATED)
+    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(['PUT'])
-def updateTasks(request):
-    tasks = task.objects.get(id=pk)
-    serializer = TaskSerializers(data=request.data)
+def updateTask(request, pk):
+    task = Task.objects.filter(id=pk).first()
+    if not task:
+        return Response({"error": "Not Found"}, status=status.HTTP_404_NOT_FOUND)
+    serializer = TaskSerializer(task, data=request.data)
     if serializer.is_valid():
         serializer.save()
-    return Response(serializer.data)
+        return Response(serializer.data, status=status.HTTP_200_OK)
+    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(['DELETE'])
-def deleteTasks(request):
-    tasks = Task.objects.get(id=pk)
+def deleteTask(request, pk):
+    task = Task.objects.filter(id=pk).first()
+    if not task:
+        return Response({"error": "Not Found"}, status=status.HTTP_404_NOT_FOUND)
     task.delete()
-    return Response("Task deleted successfully")
+    return Response({"message": "Deleted"}, status=status.HTTP_204_NO_CONTENT)
